+++ Modules/SharePointDsc/DSCResources/MSFT_SPManagedMetadataServiceApp/MSFT_SPManagedMetaDataServiceApp.psm1
@@ -77,14 +77,35 @@ function Get-TargetResource
                     $proxyName = $serviceAppProxy.Name
                 }
             }
+
+            # Get the ContentTypeHubUrl value
+            $hubUrl = ""
+            try 
+            {
+                $propertyFlags = [System.Reflection.BindingFlags]::Instance `
+                                -bor [System.Reflection.BindingFlags]::NonPublic
+
+                $propData = $serviceApp.GetType().GetMethods($propertyFlags)
+                $method = $propData | Where-Object -FilterScript {
+                    $_.Name -eq "GetContentTypeSyndicationHubLocal"
+                } 
+                $defaultPartitionId = [Guid]::Parse("0C37852B-34D0-418e-91C6-2AC25AF4BE5B")
+                $hubUrl = $method.Invoke($serviceApp, $defaultPartitionId).AbsoluteUri
+            }
+            catch [System.Exception] 
+            {
+                $hubUrl = ""
+            }
+
             return @{
-                Name            = $serviceApp.DisplayName
-                ProxyName       = $proxyName
-                Ensure          = "Present"
-                ApplicationPool = $serviceApp.ApplicationPool.Name
-                DatabaseName    = $serviceApp.Database.Name
-                DatabaseServer  = $serviceApp.Database.Server.Name
-                InstallAccount  = $params.InstallAccount
+                Name              = $serviceApp.DisplayName
+                ProxyName         = $proxyName
+                Ensure            = "Present"
+                ApplicationPool   = $serviceApp.ApplicationPool.Name
+                DatabaseName      = $serviceApp.Database.Name
+                DatabaseServer    = $serviceApp.Database.Server.Name
+                ContentTypeHubUrl = $hubUrl.TrimEnd('/')
+                InstallAccount    = $params.InstallAccount
             }
         }
     }
@@ -196,6 +217,23 @@ function Set-TargetResource
                 Set-SPMetadataServiceApplication -Identity $serviceApp -ApplicationPool $appPool
             }
         }
+
+        if (($PSBoundParameters.ContainsKey("ContentTypeHubUrl") -eq $true) `
+            -and ($ContentTypeHubUrl.TrimEnd('/') -ne $result.ContentTypeHubUrl.TrimEnd('/')))
+        {
+            Write-Verbose -Message "Updating Content type hub for Managed Metadata Service Application $Name"
+            Invoke-SPDSCCommand -Credential $InstallAccount `
+                                -Arguments $PSBoundParameters `
+                                -ScriptBlock {
+                $params = $args[0]
+                
+                $serviceApp = Get-SPServiceApplication -Name $params.Name `
+                    | Where-Object -FilterScript {
+                        $_.GetType().FullName -eq "Microsoft.SharePoint.Taxonomy.MetadataWebServiceApplication" 
+                }
+                Set-SPMetadataServiceApplication -Identity $serviceApp -HubUri $params.ContentTypeHubUrl
+            }
+        }
     }
     
     if ($Ensure -eq "Absent") 
@@ -269,12 +307,18 @@ function Test-TargetResource
     Write-Verbose -Message "Testing managed metadata service application $Name"
 
     $PSBoundParameters.Ensure = $Ensure
+    if ($PSBoundParameters.ContainsKey("ContentTypeHubUrl") -eq $true)
+    {
+        $PSBoundParameters.ContentTypeHubUrl = $ContentTypeHubUrl.TrimEnd('/')
+    }
 
     $CurrentValues = Get-TargetResource @PSBoundParameters
 
     return Test-SPDscParameterState -CurrentValues $CurrentValues `
                                     -DesiredValues $PSBoundParameters `
-                                    -ValuesToCheck @("ApplicationPool", "Ensure")
+                                    -ValuesToCheck @("ApplicationPool", 
+                                                     "ContentTypeHubUrl", 
+                                                     "Ensure")
 }
 
 Export-ModuleMember -Function *-TargetResource